refactor: swap functions' location
